More fixes for long->int32.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@13352 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/genmesh/object/genmesh.cpp
@@ -195,10 +195,10 @@ char* csGenmeshMeshObject::GenerateCacheName ()
 
   csMemFile mf;
   mf.Write ("genmesh", 7);
-  long l;
-  l = convert_endian ((long)factory->GetVertexCount ());
+  int32 l;
+  l = convert_endian ((int32)factory->GetVertexCount ());
   mf.Write ((char*)&l, 4);
-  l = convert_endian ((long)factory->GetTriangleCount ());
+  l = convert_endian ((int32)factory->GetTriangleCount ());
   mf.Write ((char*)&l, 4);
 
   if (logparent)
@@ -215,46 +215,46 @@ char* csGenmeshMeshObject::GenerateCacheName ()
         mf.Write (sect->QueryObject ()->GetName (),
 		strlen (sect->QueryObject ()->GetName ()));
       csVector3 pos = movable->GetFullPosition ();
-      l = convert_endian ((long)QInt ((pos.x * 1000)+.5));
+      l = convert_endian ((int32)QInt ((pos.x * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((pos.y * 1000)+.5));
+      l = convert_endian ((int32)QInt ((pos.y * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((pos.z * 1000)+.5));
+      l = convert_endian ((int32)QInt ((pos.z * 1000)+.5));
       mf.Write ((char*)&l, 4);
       csReversibleTransform tr = movable->GetFullTransform ();
       const csMatrix3& o2t = tr.GetO2T ();
-      l = convert_endian ((long)QInt ((o2t.m11 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m11 * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((o2t.m12 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m12 * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((o2t.m13 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m13 * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((o2t.m21 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m21 * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((o2t.m22 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m22 * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((o2t.m23 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m23 * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((o2t.m31 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m31 * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((o2t.m32 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m32 * 1000)+.5));
       mf.Write ((char*)&l, 4);
-      l = convert_endian ((long)QInt ((o2t.m33 * 1000)+.5));
+      l = convert_endian ((int32)QInt ((o2t.m33 * 1000)+.5));
       mf.Write ((char*)&l, 4);
     }
   }
 
-  l = convert_endian ((long)QInt ((b.MinX () * 1000)+.5));
+  l = convert_endian ((int32)QInt ((b.MinX () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian ((long)QInt ((b.MinY () * 1000)+.5));
+  l = convert_endian ((int32)QInt ((b.MinY () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian ((long)QInt ((b.MinZ () * 1000)+.5));
+  l = convert_endian ((int32)QInt ((b.MinZ () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian ((long)QInt ((b.MaxX () * 1000)+.5));
+  l = convert_endian ((int32)QInt ((b.MaxX () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian ((long)QInt ((b.MaxY () * 1000)+.5));
+  l = convert_endian ((int32)QInt ((b.MaxY () * 1000)+.5));
   mf.Write ((char*)&l, 4);
-  l = convert_endian ((long)QInt ((b.MaxZ () * 1000)+.5));
+  l = convert_endian ((int32)QInt ((b.MaxZ () * 1000)+.5));
   mf.Write ((char*)&l, 4);
 
   csMD5::Digest digest = csMD5::Encode (mf.GetData (), mf.GetSize ());
